requirements.txt
sanic==22.3.*
tortoise-orm==0.19.*

## development
# vim
python-lsp-server[all]
# vscode
mypy
black
